requirements.txt
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-google-genai
chromadb
sentence-transformers